Initiating with the proyect
I put the name of the proyect, prepare git for the code, battled with Love2D and VSCode setup for a while, and watched some videos to guide myself (CS50 course, first chapter: Pong)
I was inefficient, but Im going to try my best to avoid taking to much time. Its good to understand everything but the time matters.